+++ source/ui/src/main/java/com/clerk/ui/core/button/standard/ButtonStyleTokens.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.Immutable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.Dp
-import com.clerk.ui.core.common.dimens.dp0
-import com.clerk.ui.core.common.dimens.dp1
-import com.clerk.ui.core.common.dimens.dp32
-import com.clerk.ui.core.common.dimens.dp48
+import com.clerk.ui.core.dimens.dp0
+import com.clerk.ui.core.dimens.dp1
+import com.clerk.ui.core.dimens.dp32
+import com.clerk.ui.core.dimens.dp48
 import com.clerk.ui.theme.ClerkMaterialTheme
 
 @Immutable
