+++ source/ui/src/main/java/com/clerk/ui/core/input/ClerkCodeInputField.kt
@@ -58,13 +58,13 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.withStyle
 import com.clerk.ui.R
-import com.clerk.ui.core.common.dimens.dp1
-import com.clerk.ui.core.common.dimens.dp12
-import com.clerk.ui.core.common.dimens.dp24
-import com.clerk.ui.core.common.dimens.dp4
-import com.clerk.ui.core.common.dimens.dp52
-import com.clerk.ui.core.common.dimens.dp56
-import com.clerk.ui.core.common.dimens.dp8
+import com.clerk.ui.core.dimens.dp1
+import com.clerk.ui.core.dimens.dp12
+import com.clerk.ui.core.dimens.dp24
+import com.clerk.ui.core.dimens.dp4
+import com.clerk.ui.core.dimens.dp52
+import com.clerk.ui.core.dimens.dp56
+import com.clerk.ui.core.dimens.dp8
 import com.clerk.ui.signin.code.VerificationState
 import com.clerk.ui.theme.ClerkMaterialTheme
 import kotlinx.coroutines.delay
