+++ source/ui/src/main/java/com/clerk/ui/core/input/ClerkTextField.kt
@@ -33,10 +33,10 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.clerk.ui.R
-import com.clerk.ui.core.common.dimens.dp12
-import com.clerk.ui.core.common.dimens.dp20
-import com.clerk.ui.core.common.dimens.dp24
-import com.clerk.ui.core.common.dimens.dp4
+import com.clerk.ui.core.dimens.dp12
+import com.clerk.ui.core.dimens.dp20
+import com.clerk.ui.core.dimens.dp24
+import com.clerk.ui.core.dimens.dp4
 import com.clerk.ui.theme.ClerkMaterialTheme
 
 /**
