+++ source/ui/src/test/java/com/clerk/snapshot/input/ClerkPhoneNumberFieldSnapshotTest.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import com.clerk.api.Clerk
 import com.clerk.api.ui.ClerkTheme
 import com.clerk.base.BaseSnapshotTest
-import com.clerk.ui.core.common.dimens.dp12
+import com.clerk.ui.core.dimens.dp12
 import com.clerk.ui.core.input.ClerkPhoneNumberField
 import com.clerk.ui.theme.ClerkMaterialTheme
 import com.clerk.ui.theme.DefaultColors
