+++ source/ui/src/test/java/com/clerk/snapshot/input/ClerkTextFieldSnapshotTest.kt
@@ -8,8 +8,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import com.clerk.base.BaseSnapshotTest
 import com.clerk.ui.R
-import com.clerk.ui.core.common.dimens.dp12
-import com.clerk.ui.core.common.dimens.dp20
+import com.clerk.ui.core.dimens.dp12
+import com.clerk.ui.core.dimens.dp20
 import com.clerk.ui.core.input.ClerkTextField
 import com.clerk.ui.theme.ClerkMaterialTheme
 import org.junit.Test
