+++ source/ui/src/test/java/com/clerk/ui/theme/ClerkThemeSnapshotTest.kt
@@ -21,10 +21,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.clerk.api.Clerk
 import com.clerk.base.BaseSnapshotTest
-import com.clerk.ui.core.common.dimens.dp1
-import com.clerk.ui.core.common.dimens.dp12
-import com.clerk.ui.core.common.dimens.dp24
-import com.clerk.ui.core.common.dimens.dp6
+import com.clerk.ui.core.dimens.dp1
+import com.clerk.ui.core.dimens.dp12
+import com.clerk.ui.core.dimens.dp24
+import com.clerk.ui.core.dimens.dp6
 import com.materialkolor.ktx.toHex
 import org.junit.Test
